MCP/requirements.txt
openai
requests
python-dotenv
twilio
linkedin-api
